requirements.txt
asgiref==3.8.1
autopep8==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-extensions==3.2.3
django-markdownx==4.0.7
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
Faker==24.11.0
filetype==1.2.0
idna==3.7
Markdown==3.6
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
toposort==1.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
